🩹 [Patch]: Use `GitHub-Script` action (#4)
## Description

This pull request changes the action to use the `GitHub-Script` action,
to make use of built in functions in the GitHub PowerShell module:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L17-R28):
Modified the `Get-IssueFormData` step to use the
`PSModule/GitHub-Script@v1` module and updated the method for setting
the output data.
*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L117-R128):
Replaced direct output commands with `LogGroup` and `Set-GitHubOutput`
functions.
*
[`.github/workflows/Action-Test.yml`](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4L30-R43):
Replaced inline PowerShell scripts with the `PSModule/GitHub-Script@v1`
module to use the built in functionality of the GitHub PowerShell
module.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L84-R84):
Updated the `Get-IssueFormData` module from version `v0` to `v1`.


## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas