🩹 [Patch]: Make outputs not break line (#10)
## Description

This pull request includes changes to the `scripts/main.ps1` file to
improve the formatting and output rendering of the script. The most
important changes include setting the output rendering style to 'Ansi'
and ensuring the formatted list is converted to a string before being
output.

Improvements to output formatting:

* `scripts/main.ps1`:
* Set `$PSStyle.OutputRendering` to 'Ansi' to ensure proper output
rendering.
* Modified the pipeline to convert the formatted list to a string using
`Out-String` before outputting it.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas