🪲 [Fix]: Disable token on the script (#11)
## Description

This pull request includes minor updates to the GitHub Action by
disabling the logon functionality by providing an empty token.

Updates to GitHub Actions configuration:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R28):
Added a default empty value for the `Token` parameter under `with:` to
ensure that the action does not require.

Simplifications to PowerShell script:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L8-L9):
Removed `$PSStyle.OutputRendering = 'Ansi'` to simplify the script and
potentially avoid unnecessary styling configuration. This is included in
the `GitHub` PowerShell module used via `GitHub-Script`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas